requirements.txt
Django==3.0.14
djangorestframework==3.11.0
psycopg2-binary==2.8.6
drf-yasg==1.20.0
django-redis==5.0.0

# developer
flake8==3.9.2
django-extensions==3.1.3
vulture==2.3
mock==4.0.3
